Introduce Lens to allow for conditional updates
The Lens is used to modify the state into the shape that the component
requires, which allows Reductive.Provider to implement shouldUpdate(). This
avoids needless updates when another part of the application state changes.
You can observe the change using React devtools "highlight updates."

Introduces a new type, Lens, and all users of Provider must now provide an
implementation of this type as an argument to createMake(). Wanted to make
this argument optional, but the default would return a 'state instead of a
'lensed, and I didn't know how to resolve that.

This change includes a new example specifically constructed to demonstrate
the problem, or lack thereof now that Lens is in use by Provider.