Upgrade all dependencies
Because the dependencies are very old, horrifyingly confusing error messages
occur at run time within any application that depends on reductive and uses
a more recent version of ReasonReact. In particular, the newer ReasonReact
in the application cosntructs a component whose shape differs from the
version of ReasonReact in reductive, leading to a bizarre error in the curry
module within BuckleScript.

Also fix the webpack config file formatting.