Move lens to a dedicated component
Using code from Taras Vozniuk, the Provider component now delegates to a new
Lense component, which contains the bulk of the original Provider code. The
"new" Provider implements a default identity function for the lense.

This approach retains the desired behavior of rendering only the changed
parts of the application, as well as remaining backward compatible.

Seems that refmt changed some parentheses to curly braces.